Change text and create product array with first product item
Added array with first product item object which included name,
description, id and image imported from asset. This is added to state
to allow the use of the object in the store component.
Changed homepage text to better reflect store's function. Edit alt of
product image to represent the name of the product item object.