Add and edit logic to open and close basket and add items to the basket
Linked product id with id in the button for product in shop to allow
for item to be added to the basket object. Changed open basket state
to be a property of basket object and edited logic accordingly. Added
basket object and logic for adding items from the product array to the
basket object to allow it to be displayed in the checkout component.