Solve mutability issues, restore add and subtract functionality
Use mapping to allow for increments and decrements of items in basket
without mutating state. Before resolving this the code was adding two
of the same item at a time. Reorganise code to create two arrays of
totalPrice and shopping cart to allow for easier manipulation of these
in state.